refactor(parser): revert panic on expect from gofuzz
This didn't work as I expected it to. I have been working with gofuzz
and this change makes it impossible to fuzz other smaller incremental
changes without hitting an overly strict expect.

This concept may be used in the future, but it will have to first turn
this functionality into two separate methods were one is an assertion
and one skips tokens and records errors like the current implementation.

This reverts commit fa4ea564956ea489ade8d9117e1f1830c387997e.
